Semi-finished with profile and index
The carrots on the nav bars took a very long time to create.  However,
I was able to create them and have active classes accordingly to style
the highlighted nav bar.  The only issue I have is with the icon color
images.  However, I believe you can change the css of the svg files
with the fill property.  I will have to ask Won or Wolfram regarding
this matter. Anyways I just wanted to upload this to make sure I have a
backup version of what I have so far….